Clarify dry-run results vs actual copy
I found that running with the `--dry-run` option gives the same output as running without it, which can be confusing. Example:

INFO: copying /prod/db/secret to /qa/db/secret
INFO: copying /prod/db/table to /qa/db/table

The tool gives the impression that it has copied parameters, when in fact it has not because it was just a dry run.

Also, made a minor change to redirect this output to `stdout` rather than `stderr`, as it is not an error.

Very helpful project, thank you!